backend/requirements.txt
# Core Django
Django==5.2.8
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.9.0
python-decouple==3.8
drf-yasg==1.21.11

# Database
psycopg2-binary==2.9.11
django-environ==0.11.2

# Async & HTTP
httpx==0.26.0  # python-telegram-bot bilan mos kelishi uchun 0.26.0 versiyasiga o'zgartirildi
aiohttp==3.9.5
requests==2.32.5

# Telegram Bot
aiogram==3.22.0
python-telegram-bot==20.8

# Data Processing
pandas==2.3.3
openpyxl==3.1.5
python-dateutil==2.9.0

# Production
gunicorn==21.2.0
whitenoise==6.7.0
uvicorn==0.29.0

# Utils
python-dotenv==1.2.1
python-multipart==0.0.9
loguru==0.7.2